Use config for websocket security key
Previously this had a hardcoded value.
I'm uncertain what this key is needed for, but I didn't want to
make a bunch of local changes to this library.
This might not have any effect, but just in case, it is a best
practice to move it outside of the codebase.