Fix delay and mix routing
Fix issue with delay routing. Previously, the order of the nodes
within the group were uncontrolled, and new nodes always went to
the tail of the group. This change moves new nodes to head after
declaration. The effect is that the signals are sent to the bus
before the delay node reads it, so the bus is no longer silent.

Also add "master" mix Ndef with gain (pre-tanh) and vol controls.
This of course is after the delay.
(still needs lua mapping)