Add mix_gain, mix_vol engine controls
The mix_gain control has no ceiling. The overall mix is multiplied
by this before being soft-clipped via .tanh

The mix_vol control is restricted between 0 and 1. The post-tanh
signal is multiplied by this.

It might be good to add a bit of lag to these to smooth them out.